aun mas cambios, u.u

+++ sources/RuntimeEditor/Windows/RuntimeEditorWindowMaterial.cs
@@ -54,7 +54,7 @@ public partial class RuntimeEditorWindowMaterial : Window
     private void CargarMods()
     {
         var listMods = TableMods.Instance.ObtenerTodos().ToList();
-        foreach (System.Collections.Generic.KeyValuePair<byte, ModInfo> item in listMods)
+        foreach (System.Collections.Generic.KeyValuePair<ushort, ModInfo> item in listMods)
         {
             ListaTipos.AddItemWithData(item.Value.Name, item.Value);
         }
